+++ frontend/package.json
@@ -26,7 +26,6 @@
     "@fluentui/react-icons": "^2.0.258",
     "axios": "^1.13.5",
     "react": "^18.3.1",
-    "react-async-hook": "^4.0.0",
     "react-dom": "^18.3.1",
     "react-error-boundary": "^6.1.1"
   },
