+++ frontend/src/components/Chat/ConverterPanel/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ConverterPanel'
